Add variables arduino

+++ src/obstacle_avoidance/src/obstacle_avoidance_node.cpp
@@ -169,7 +169,7 @@ class ObstacleAvoidance {
     local_goal_angle =
         std::min(std::max(local_goal_angle, float(-90.0)), float(90.0));
     int throttle = int(local_goal_speed * THROTTLE_FORWARD);
-    int servo = SERVO_LEFT * local_goal_angle / 90.0;
+    int servo = int(SERVO_LEFT * local_goal_angle / 90.0);
     std::printf("final_goal_speed: %.2f\n", local_goal_speed);
     std::printf("final_goal_angle: %.2f\n", local_goal_angle);
     if (emergency_stop) {
